+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Breaking
 
 ### Changed
-
+-  Remove unused fields from TimeUtil [#442](https://github.com/ie3-institute/PowerSystemUtils/issues/442)
 
 ## [2.1.0]